filter module + card module were added
